Revert "test: restructure test files"
This reverts commit 29432e7153c6f10976d8964a81e4f3d89037f36e.
